refactor: 7707 remove Shared.API and move code to respective projects
