+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.API/xxENSONOxx.xxSTACKSxx.API.csproj
@@ -12,7 +12,6 @@
     <ItemGroup>
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.CQRS\xxENSONOxx.xxSTACKSxx.CQRS.csproj" />
         <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Infrastructure\xxENSONOxx.xxSTACKSxx.Infrastructure.csproj" />
-        <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.API\xxENSONOxx.xxSTACKSxx.Shared.API.csproj" />
         <ProjectReference Include="..\..\..\..\shared\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions\xxENSONOxx.xxSTACKSxx.Shared.Application.CQRS.Abstractions.csproj" />
     </ItemGroup>
