--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.API.Tests/Usings.cs
@@ -1,4 +1,0 @@
-global using AutoFixture.Xunit2;
-global using FluentAssertions;
-global using TestStack.BDDfy;
-global using Xunit;
