--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.API/Configuration/CorrelationIdConfiguration.cs
@@ -1,11 +1,0 @@
-namespace xxENSONOxx.xxSTACKSxx.Shared.API.Configuration
-{
-    public class CorrelationIdConfiguration
-    {
-        private const string DefaultHeaderName = "x-correlation-id";
-
-        public string HeaderName { get; set; } = DefaultHeaderName;
-
-        public bool IncludeInResponse { get; set; } = true;
-    }
-}
