--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.API/xxENSONOxx.xxSTACKSxx.Shared.API.csproj
@@ -1,28 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Description>API package for common features used throughout the projects.</Description>
-    <Company>Ensono</Company>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/ensono/stacks-dotnet-packages-api</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/ensono/stacks-dotnet-packages-api</RepositoryUrl>
-    <RepositoryType>GIT</RepositoryType>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="Serilog" Version="4.0.1" />
-    <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
-    <PackageReference Include="System.Text.Encodings.Web" Version="8.0.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\xxENSONOxx.xxSTACKSxx.Shared.Core\xxENSONOxx.xxSTACKSxx.Shared.Core.csproj" />
-  </ItemGroup>
-
-</Project>
