+++ src/simple-api/.template.config/template.json
@@ -116,9 +116,6 @@
 				".gitignore",
 				".gitattributes",
 				"src/simple-api/**",
-				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.API/**",
-				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.API.Swagger/**",
-				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.API.Tests/**",
 				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.Core/**",
 				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.Core.Tests/**",
 				"src/shared/xxENSONOxx.xxSTACKSxx.Shared.Testing/**",
