+++ src/simple-api/src/api/xxENSONOxx.xxSTACKSxx.API/Configuration/CorrelationIdConfiguration.cs
@@ -1,0 +1,10 @@
+namespace xxENSONOxx.xxSTACKSxx.API.Configuration;
+
+public class CorrelationIdConfiguration
+{
+    private const string DefaultHeaderName = "x-correlation-id";
+
+    public string HeaderName { get; set; } = DefaultHeaderName;
+
+    public bool IncludeInResponse { get; set; } = true;
+}
