+++ src/simple-api/src/api/xxENSONOxx.xxSTACKSxx.API/Middleware/CorrelationIdMiddlewareExtensions.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Builder;
+
+namespace xxENSONOxx.xxSTACKSxx.API.Middleware;
+
+public static class CorrelationIdMiddlewareExtensions
+{
+    public static void UseCorrelationId(this IApplicationBuilder app)
+    {
+        app.UseMiddleware<CorrelationIdMiddleware>();
+    }
+}
