+++ xxENSONOxx.xxSTACKSxx.sln
@@ -37,10 +37,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.API.U
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Infrastructure.UnitTests", "src\cqrs\src\api\xxENSONOxx.xxSTACKSxx.Infrastructure.UnitTests\xxENSONOxx.xxSTACKSxx.Infrastructure.UnitTests.csproj", "{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.API", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.API\xxENSONOxx.xxSTACKSxx.Shared.API.csproj", "{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.API.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.API.Tests\xxENSONOxx.xxSTACKSxx.Shared.API.Tests.csproj", "{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Testing", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Testing\xxENSONOxx.xxSTACKSxx.Shared.Testing.csproj", "{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests", "src\shared\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests\xxENSONOxx.xxSTACKSxx.Shared.Testing.Tests.csproj", "{6D545555-2426-4478-82A9-D756A00F1E39}"
@@ -325,30 +321,6 @@ Global
 		{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3}.Release|x64.Build.0 = Release|Any CPU
 		{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3}.Release|x86.ActiveCfg = Release|Any CPU
 		{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3}.Release|x86.Build.0 = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|x64.Build.0 = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Debug|x86.Build.0 = Debug|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|Any CPU.Build.0 = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|x64.ActiveCfg = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|x64.Build.0 = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|x86.ActiveCfg = Release|Any CPU
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72}.Release|x86.Build.0 = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|x64.Build.0 = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Debug|x86.Build.0 = Debug|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x64.ActiveCfg = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x64.Build.0 = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x86.ActiveCfg = Release|Any CPU
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C}.Release|x86.Build.0 = Release|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -778,8 +750,6 @@ Global
 		{EEDF6DBD-2E45-4E41-B027-B88F8293C56F} = {A93ABC7D-17EC-4D7E-B8CF-94F11D94910D}
 		{ADA636E5-76CF-4B43-87C2-F53E37150A8B} = {A93ABC7D-17EC-4D7E-B8CF-94F11D94910D}
 		{BEE416A3-8EB2-4415-9518-0D0D58EDA7B3} = {A93ABC7D-17EC-4D7E-B8CF-94F11D94910D}
-		{78EA9F72-72DA-4F0C-A647-B315CF9DBD72} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
-		{F4B072C7-1AF9-4ADE-A055-7C0E4257035C} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{5375FD38-BB3F-45B6-A42C-E62D7F4C3834} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{6D545555-2426-4478-82A9-D756A00F1E39} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
 		{DAA32E63-80B5-49B6-890E-7A11FDA6A85B} = {5C2EF365-4E4D-4446-A3C6-325168F0531F}
